src/api/requirements.txt
# =========================
# FoxProFlow — requirements.txt (single source of truth)
# Построен на базе действующего requirements.txt; файл requirements_new.txt удаляем.
# =========================

# --- Web / API ---
fastapi==0.115.13
starlette==0.46.2
uvicorn==0.34.3
orjson==3.11.4

# --- Validation / typing ---
pydantic==2.11.7
pydantic_core==2.33.2
typing_extensions==4.14.1
typing-inspection==0.4.1
annotated-types==0.7.0

# --- DB / SQL ---
SQLAlchemy==2.0.40
psycopg[binary]==3.1.18         # psycopg3 (основной драйвер)
psycopg2-binary==2.9.9          # совместимость с кодом/роутами на psycopg2
asyncpg==0.30.0                 # единая версия (в new.txt была 0.29.0 — не используем)
greenlet==3.2.4

# --- Redis / Celery ---
redis==5.2.1
celery==5.4.0

# --- HTTP / Network ---
requests==2.32.4
aiohttp==3.9.5
httpx==0.27.2 ; python_version >= "3.8"   # OSRM/Valhalla-клиент
h11==0.16.0
anyio==4.10.0
sniffio==1.3.1
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
certifi==2025.8.3
PySocks==1.7.1
websocket-client==1.8.0
wsproto==1.2.0

# --- aiohttp deps ---
aiosignal==1.4.0
async-timeout==5.0.1
frozenlist==1.7.0
multidict==6.6.4
yarl==1.20.1
attrs==25.3.0

# --- HTML / Parsing ---
beautifulsoup4==4.12.3
soupsieve==2.8
lxml==5.4.0
PyYAML==6.0.2

# --- Data / Scientific ---
numpy==2.2.6
pandas==2.2.3
scipy==1.14.1
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# --- Selenium (если парсеры браузерные) ---
selenium==4.35.0
webdriver-manager==4.0.0
outcome==1.3.0.post0
trio==0.30.0
trio-websocket==0.12.2
sortedcontainers==2.4.0

# --- Ops / Utils ---
packaging==25.0
python-dotenv==1.1.1
colorama==0.4.6
click==8.2.1
psutil==7.0.0
propcache==0.3.2
cffi==1.17.1
pycparser==2.22
psycopg[binary]>=3.1
psycopg_pool>=3.2
